refactor(project): refactor to separated modules
issue#2
