+++ src/ng-module-registration.ts
@@ -1,0 +1,7 @@
+import { restangularProvider } from './restangular-provider';
+
+const restangular = angular.module('restangular', []);
+
+restangular.provider('Restangular', restangularProvider);
+
+export default restangular.name;
