+++ tsconfig.json
@@ -6,11 +6,9 @@
     "target": "es5",
     "allowJs": true,
     "sourceMap": true,
-    "types": [
-			"angular",
-			"angular-mocks",
-			"jasmine"
-    ]
+    "esModuleInterop": true,
+    "allowSyntheticDefaultImports": true,
+    "moduleResolution": "node"
   },
   "exclude": [
 		"out",
